requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.14,<3.15
django-cors-headers>=3.13,<4.0

